iOS app upload.
To use the iOS app, create and sign in to Xcode with your Apple Developer account and run it on your iOS device. Grant permissions for using Healthkit data, and remember to whitelist your server address in Signing & Capabilities. Run the iOS app once the server is up and running.